Add content script for DOM inspection.
The content script scans the DOM for positives and sends it to the
browserAction extension. If none is detected, the browserAction is
disabled for this tab.id. If positives are detected, the browserAction
gets a badge with positives.length and clicking the browserAction
inserts the positives into the DOM.

Next up: input type detection!